avoid checking for import methods in parent classes
Since we're checking for a import method getting installed, it makes
sense to check for it directly in the namespace and ignoring parent
classes.

Future versions of perl are intending to have a real UNIVERSAL::import
method rather than a magic special case. Ignoring parent classes when
checking for an installed import avoids finding UNIVERSAL::import.